requirements.txt
azure-identity
azure-storage-blob
sqlalchemy
pyodbc
pandas
python-dotenv
